app_client/requirements.txt
streamlit
langchain-google-genai
langchain-mcp-adapters
langgraph
python-dotenv